Handle RubyGems returning invalid json
If you use non-RubyGems sources, you may find that querying RubyGems
results in a non-JSON response. This adds a test case covering that and
handling it similar to other failure modes.

This also introduces WebMock in order to make the tests more-agnostic as
to how RubyGems gets queried.

Last thing, this changes string-interpolation for RubyGems urls to use
Addressable URI templates since it is both convenient for testing, but
also a better representation for how the constant is used.